epaint

Enum Shape

pub enum Shape {
    Noop,
    Vec(Vec<Self>),
    Circle(CircleShape),
    Ellipse(EllipseShape),
    LineSegment {
        points: [Pos2; 2],
        stroke: Stroke,
    },
    Path(PathShape),
    Rect(RectShape),
    Text(TextShape),
    Mesh(Arc<Mesh>),
    QuadraticBezier(QuadraticBezierShape),
    CubicBezier(CubicBezierShape),
    Callback(PaintCallback),
}

A paint primitive such as a circle or a piece of text. Coordinates are all screen space points (not physical pixels).

You should generally recreate your Shapes each frame, but storing them should also be fine with one exception: Shape::Text depends on the current pixels_per_point (dpi scale) and so must be recreated every time pixels_per_point changes.

Variants

Noop

Paint nothing. This can be useful as a placeholder.

Vec(Vec<Self>)

Recursively nest more shapes - sometimes a convenience to be able to do. For performance reasons it is better to avoid it.

Circle(CircleShape)

Circle with optional outline and fill.

Ellipse(EllipseShape)

Ellipse with optional outline and fill.

LineSegment

A line between two points.

Fields

points: [Pos2; 2]

stroke: Stroke

Path(PathShape)

A series of lines between points. The path can have a stroke and/or fill (if closed).

Rect(RectShape)

Rectangle with optional outline and fill.

Text(TextShape)

Text.

This needs to be recreated if pixels_per_point (dpi scale) changes.

Mesh(Arc<Mesh>)

A general triangle mesh.

Can be used to display images.

Wrapped in an Arc to minimize the size of Shape.

QuadraticBezier(QuadraticBezierShape)

A quadratic Bézier Curve.

CubicBezier(CubicBezierShape)

A cubic Bézier Curve.

Callback(PaintCallback)

Backend-specific painting.

Implementations

impl Shape

Constructors

pub fn line_segment(points: [Pos2; 2], stroke: impl Into<Stroke>) -> Self

A line between two points. More efficient than calling Self::line.

pub fn hline(x: impl Into<Rangef>, y: f32, stroke: impl Into<Stroke>) -> Self

A horizontal line.

pub fn vline(x: f32, y: impl Into<Rangef>, stroke: impl Into<Stroke>) -> Self

A vertical line.

pub fn line(points: Vec<Pos2>, stroke: impl Into<PathStroke>) -> Self

A line through many points.

Use Self::line_segment instead if your line only connects two points.

pub fn closed_line(points: Vec<Pos2>, stroke: impl Into<PathStroke>) -> Self

A line that closes back to the start point again.

pub fn dotted_line( path: &[Pos2], color: impl Into<Color32>, spacing: f32, radius: f32, ) -> Vec<Self>

Turn a line into equally spaced dots.

pub fn dashed_line( path: &[Pos2], stroke: impl Into<Stroke>, dash_length: f32, gap_length: f32, ) -> Vec<Self>

Turn a line into dashes.

pub fn dashed_line_with_offset( path: &[Pos2], stroke: impl Into<Stroke>, dash_lengths: &[f32], gap_lengths: &[f32], dash_offset: f32, ) -> Vec<Self>

Turn a line into dashes with different dash/gap lengths and a start offset.

pub fn dashed_line_many( points: &[Pos2], stroke: impl Into<Stroke>, dash_length: f32, gap_length: f32, shapes: &mut Vec<Self>, )

Turn a line into dashes. If you need to create many dashed lines use this instead of Self::dashed_line.

pub fn dashed_line_many_with_offset( points: &[Pos2], stroke: impl Into<Stroke>, dash_lengths: &[f32], gap_lengths: &[f32], dash_offset: f32, shapes: &mut Vec<Self>, )

Turn a line into dashes with different dash/gap lengths and a start offset. If you need to create many dashed lines use this instead of Self::dashed_line_with_offset.

pub fn convex_polygon( points: Vec<Pos2>, fill: impl Into<Color32>, stroke: impl Into<PathStroke>, ) -> Self

A convex polygon with a fill and optional stroke.

The most performant winding order is clockwise.

pub fn circle_filled( center: Pos2, radius: f32, fill_color: impl Into<Color32>, ) -> Self

pub fn circle_stroke( center: Pos2, radius: f32, stroke: impl Into<Stroke>, ) -> Self

pub fn ellipse_filled( center: Pos2, radius: Vec2, fill_color: impl Into<Color32>, ) -> Self

pub fn ellipse_stroke( center: Pos2, radius: Vec2, stroke: impl Into<Stroke>, ) -> Self

pub fn rect_filled( rect: Rect, corner_radius: impl Into<CornerRadius>, fill_color: impl Into<Color32>, ) -> Self

See also Self::rect_stroke.

pub fn rect_stroke( rect: Rect, corner_radius: impl Into<CornerRadius>, stroke: impl Into<Stroke>, stroke_kind: StrokeKind, ) -> Self

See also Self::rect_filled.

pub fn gradient_rect( rect: Rect, direction: Direction, [from, to]: [Color32; 2], ) -> Self

Paints a gradient rectangle that transitions from color_from to color_to along the given direction.

For example, Direction::TopDown paints color_from at the top edge fading to color_to at the bottom edge.

pub fn text( fonts: &mut FontsView<'_>, pos: Pos2, anchor: Align2, text: impl ToString, font_id: FontId, color: Color32, ) -> Self

pub fn galley(pos: Pos2, galley: Arc<Galley>, fallback_color: Color32) -> Self

Any uncolored parts of the Galley (using Color32::PLACEHOLDER) will be replaced with the given color.

Any non-placeholder color in the galley takes precedence over this fallback color.

pub fn galley_with_override_text_color( pos: Pos2, galley: Arc<Galley>, text_color: Color32, ) -> Self

All text color in the Galley will be replaced with the given color.

pub fn galley_with_color( pos: Pos2, galley: Arc<Galley>, text_color: Color32, ) -> Self

👎Deprecated:
Use Shape::galley or Shape::galley_with_override_text_color instead

pub fn mesh(mesh: impl Into<Arc<Mesh>>) -> Self

pub fn image(texture_id: TextureId, rect: Rect, uv: Rect, tint: Color32) -> Self

An image at the given position.

uv should normally be Rect::from_min_max(pos2(0.0, 0.0), pos2(1.0, 1.0)) unless you want to crop or flip the image.

tint is a color multiplier. Use Color32::WHITE if you don’t want to tint the image.

pub fn visual_bounding_rect(&self) -> Rect

The visual bounding rectangle (includes stroke widths)

impl Shape

Inspection and transforms

pub fn texture_id(&self) -> TextureId

pub fn scale(&mut self, factor: f32)

Scale the shape by factor, in-place.

A wrapper around Self::transform.

pub fn translate(&mut self, delta: Vec2)

Move the shape by delta, in-place.

A wrapper around Self::transform.

pub fn transform(&mut self, transform: TSTransform)

Transform (move/scale) the shape in-place.

If using a PaintCallback, note that only the rect is scaled as opposed to other shapes where the stroke is also scaled.