epaint::image
Struct FontImage

pub struct FontImage {
    pub size: [usize; 2],
    pub pixels: Vec<f32>,
}

A single-channel image designed for the font texture.

Each value represents โcoverageโ, i.e. how much a texel is covered by a character.

This is roughly interpreted as the opacity of a white image.

Fields
size: [usize; 2]

width, height

pixels: Vec<f32>

The coverage value.

Often you want to use Self::srgba_pixels instead.

Implementations

impl FontImage

pub fn new(size: [usize; 2]) -> Self

pub fn width(&self) -> usize

pub fn height(&self) -> usize

pub fn srgba_pixels( &self, gamma: Option<f32>, ) -> impl ExactSizeIterator<Item = Color32> + '_

Returns the textures as sRGBA premultiplied pixels, row by row, top to bottom.

gamma should normally be set to None.

If you are having problems with text looking skinny and pixelated, try using a low gamma, e.g. 0.4.

pub fn region(&self, [x, y]: [usize; 2], [w, h]: [usize; 2]) -> Self

Clone a sub-region as a new image.