epaint::margin

Struct Margin

pub struct Margin {
    pub left: i8,
    pub right: i8,
    pub top: i8,
    pub bottom: i8,
}

A value for all four sides of a rectangle, often used to express padding or spacing.

Can be added and subtracted to/from Rects.

Negative margins are possible, but may produce weird behavior. Use with care.

All values are stored as i8 to keep the size of Margin small. If you want floats, use crate::MarginF32 instead.

Fields
left: i8right: i8top: i8bottom: i8
Implementations

impl Margin

pub const ZERO: Self

pub const fn same(margin: i8) -> Self

The same margin on every side.

pub const fn symmetric(x: i8, y: i8) -> Self

Margins with the same size on opposing sides

pub const fn leftf(self) -> f32

Left margin, as f32

pub const fn rightf(self) -> f32

Right margin, as f32

pub const fn topf(self) -> f32

Top margin, as f32

pub const fn bottomf(self) -> f32

Bottom margin, as f32

pub fn sum(self) -> Vec2

Total margins on both sides

pub const fn left_top(self) -> Vec2

pub const fn right_bottom(self) -> Vec2

pub const fn is_same(self) -> bool

Are the margin on every side the same?