Module epaint::mutex

Helper module that wraps some Mutex types with different implementations.

When the single_threaded feature is on the mutexes will panic when locked from different threads.

Structs

Mutex

Provides interior mutability. Only thread-safe if the multi_threaded feature is enabled.

MutexGuard

The lock you get from Mutex.

RwLock

Provides interior mutability. Only thread-safe if the multi_threaded feature is enabled.

RwLockReadGuard

The lock you get from RwLock::read.

RwLockWriteGuard

The lock you get from RwLock::write.