epaint

Module shape

The different shapes that can be painted.

Re-exports

pub use crate::CubicBezierShape;

pub use crate::QuadraticBezierShape;

Structs

CircleShape

How to paint a circle.

EllipseShape

How to paint an ellipse.

PaintCallback

If you want to paint some 3D shapes inside an egui region, you can use this.

PaintCallbackInfo

Information passed along with PaintCallback (Shape::Callback).

PathShape

A path which can be stroked and/or filled (if closed).

RectShape

How to paint a rectangle.

Rounding

How rounded the corners of things should be

TextShape

How to paint some text on screen.

ViewportInPixels

Size of the viewport in whole, physical pixels.

Enums

Shape

A paint primitive such as a circle or a piece of text. Coordinates are all screen space points (not physical pixels).

Functions

dashes_from_line 🔒

Creates dashes from a line.

points_from_line 🔒

Creates equally spaced filled circles from a line.