epaint

Struct CornerRadius

pub struct CornerRadius {
    pub nw: u8,
    pub ne: u8,
    pub sw: u8,
    pub se: u8,
}

How rounded the corners of things should be.

This specific the corner radius of the underlying geometric shape (e.g. rectangle). If there is a stroke, then the stroke will have an inner and outer corner radius which will depends on its width and crate::StrokeKind.

The rounding uses u8 to save space, so the amount of rounding is limited to integers in the range [0, 255].

For calculations, you may want to use crate::CornerRadiusF32 instead, which uses f32.

Fields
nw: u8

Radius of the rounding of the North-West (left top) corner.

ne: u8

Radius of the rounding of the North-East (right top) corner.

sw: u8

Radius of the rounding of the South-West (left bottom) corner.

se: u8

Radius of the rounding of the South-East (right bottom) corner.

Implementations

impl CornerRadius

pub const ZERO: Self

No rounding on any corner.

pub const fn same(radius: u8) -> Self

Same rounding on all four corners.

pub fn is_same(self) -> bool

Do all corners have the same rounding?

pub fn at_least(self, min: u8) -> Self

Make sure each corner has a rounding of at least this.

pub fn at_most(self, max: u8) -> Self

Make sure each corner has a rounding of at most this.

pub fn average(&self) -> f32

Average rounding of the corners.