Struct epaint::Mesh

pub struct Mesh {
    pub indices: Vec<u32>,
    pub vertices: Vec<Vertex>,
    pub texture_id: TextureId,
}

Textured triangles in two dimensions.

Fields
indices: Vec<u32>

Draw as triangles (i.e. the length is always multiple of three).

If you only support 16-bit indices you can use Mesh::split_to_u16.

egui is NOT consistent with what winding order it uses, so turn off backface culling.

vertices: Vec<Vertex>

The vertex data indexed by indices.

texture_id: TextureId

The texture to use when drawing these triangles.

Implementations

impl Mesh

pub fn with_texture(texture_id: TextureId) -> Self

pub fn clear(&mut self)

Restore to default state, but without freeing memory.

pub fn bytes_used(&self) -> usize

pub fn is_valid(&self) -> bool

Are all indices within the bounds of the contained vertices?

pub fn is_empty(&self) -> bool

pub fn calc_bounds(&self) -> Rect

Calculate a bounding rectangle.

pub fn append(&mut self, other: Self)

Append all the indices and vertices of other to self.

pub fn append_ref(&mut self, other: &Self)

Append all the indices and vertices of other to self without taking ownership.

pub fn colored_vertex(&mut self, pos: Pos2, color: Color32)

pub fn add_triangle(&mut self, a: u32, b: u32, c: u32)

Add a triangle.

pub fn reserve_triangles(&mut self, additional_triangles: usize)

Make room for this many additional triangles (will reserve 3x as many indices). See also reserve_vertices.

pub fn reserve_vertices(&mut self, additional: usize)

Make room for this many additional vertices. See also reserve_triangles.

pub fn add_rect_with_uv(&mut self, rect: Rect, uv: Rect, color: Color32)

Rectangle with a texture and color.

pub fn add_colored_rect(&mut self, rect: Rect, color: Color32)

Uniformly colored rectangle.

pub fn split_to_u16(self) -> Vec<Mesh16>

This is for platforms that only support 16-bit index buffers.

Splits this mesh into many smaller meshes (if needed) where the smaller meshes have 16-bit indices.

pub fn translate(&mut self, delta: Vec2)

Translate location by this much, in-place

pub fn transform(&mut self, transform: TSTransform)

Transform the mesh in-place with the given transform.

pub fn rotate(&mut self, rot: Rot2, origin: Pos2)

Rotate by some angle about an origin, in-place.

Origin is a position in screen space.