epaint

Struct TextureAtlas

pub struct TextureAtlas {
    image: ColorImage,
    dirty: Rectu,
    cursor: (usize, usize),
    row_height: usize,
    overflowed: bool,
    discs: Vec<PrerasterizedDisc>,
    options: TextOptions,
}

Contains font data in an atlas, where each character occupied a small rectangle.

More characters can be added, possibly expanding the texture.

Fields
image: ColorImagedirty: Rectu

What part of the image that is dirty

cursor: (usize, usize)

Used for when allocating new rectangles.

row_height: usizeoverflowed: bool

Set when someone requested more space than was available.

discs: Vec<PrerasterizedDisc>

pre-rasterized discs of radii 2^i, where i is the index.

options: TextOptions

Controls how to convert glyph coverage to alpha.

Implementations

impl TextureAtlas

pub fn new(size: [usize; 2], options: TextOptions) -> Self

pub fn options(&self) -> &TextOptions

pub fn size(&self) -> [usize; 2]

pub fn prepared_discs(&self) -> Vec<PreparedDisc>

Returns the locations and sizes of pre-rasterized discs (filled circles) in this atlas.

fn max_height(&self) -> usize

pub fn fill_ratio(&self) -> f32

When this get high, it might be time to clear and start over!

pub fn texture_options() -> TextureOptions

The texture options suitable for a font texture

pub fn image(&self) -> &ColorImage

The full font atlas image.

pub fn take_delta(&mut self) -> Option<ImageDelta>

Call to get the change to the image since last call.

pub fn allocate( &mut self, (w, h): (usize, usize), ) -> ((usize, usize), &mut ColorImage)

Returns the coordinates of where the rect ended up, and invalidates the region.