Struct epaint::TextureHandle

pub struct TextureHandle { /* private fields */ }

Used to paint images.

An image is pixels stored in RAM, and represented using ImageData. Before you can paint it however, you need to convert it to a texture.

If you are using egui, use egui::Context::load_texture.

The TextureHandle can be cloned cheaply. When the last TextureHandle for specific texture is dropped, the texture is freed.

See also TextureManager.

Implementations

impl TextureHandle

pub fn new(tex_mngr: Arc<RwLock<TextureManager>>, id: TextureId) -> Self

If you are using egui, use egui::Context::load_texture instead.

pub fn id(&self) -> TextureId

pub fn set(&mut self, image: impl Into<ImageData>, options: TextureOptions)

Assign a new image to an existing texture.

pub fn set_partial( &mut self, pos: [usize; 2], image: impl Into<ImageData>, options: TextureOptions )

Assign a new image to a subregion of the whole texture.

pub fn size(&self) -> [usize; 2]

width x height

pub fn size_vec2(&self) -> Vec2

width x height

pub fn byte_size(&self) -> usize

width x height x bytes_per_pixel

pub fn aspect_ratio(&self) -> f32

width / height

pub fn name(&self) -> String

Debug-name.