epaint::text::font

Struct FontCell

#[repr(transparent)]
struct FontCell {
    unsafe_self_cell: UnsafeSelfCell<FontCell, Arc<dyn AsRef<[u8]> + Send + Sync>, DependentFontData<'static>>,
}

Fields
unsafe_self_cell: UnsafeSelfCell<FontCell, Arc<dyn AsRef<[u8]> + Send + Sync>, DependentFontData<'static>>
Implementations

impl FontCell

fn new( owner: Arc<dyn AsRef<[u8]> + Send + Sync>, dependent_builder: impl for<'_q> FnOnce(&'_q Arc<dyn AsRef<[u8]> + Send + Sync>) -> DependentFontData<'_q>, ) -> Self

Constructs a new self-referential struct.

The provided owner will be moved into a heap allocated box. Followed by construction of the dependent value, by calling dependent_builder with a shared reference to the owner that remains valid for the lifetime of the constructed struct.

fn try_new<Err>( owner: Arc<dyn AsRef<[u8]> + Send + Sync>, dependent_builder: impl for<'_q> FnOnce(&'_q Arc<dyn AsRef<[u8]> + Send + Sync>) -> Result<DependentFontData<'_q>, Err>, ) -> Result<Self, Err>

Constructs a new self-referential struct or returns an error.

Consumes owner on error.

fn try_new_or_recover<Err>( owner: Arc<dyn AsRef<[u8]> + Send + Sync>, dependent_builder: impl for<'_q> FnOnce(&'_q Arc<dyn AsRef<[u8]> + Send + Sync>) -> Result<DependentFontData<'_q>, Err>, ) -> Result<Self, (Arc<dyn AsRef<[u8]> + Send + Sync>, Err)>

Constructs a new self-referential struct or returns an error.

Returns owner and error as tuple on error.

fn borrow_owner<'_q>(&'_q self) -> &'_q Arc<dyn AsRef<[u8]> + Send + Sync>

Borrows owner.

fn with_dependent<'outer_fn, Ret>( &'outer_fn self, func: impl for<'_q> FnOnce(&'_q Arc<dyn AsRef<[u8]> + Send + Sync>, &'outer_fn DependentFontData<'_q>) -> Ret, ) -> Ret

Calls given closure func with a shared reference to dependent.

fn with_dependent_mut<'outer_fn, Ret>( &'outer_fn mut self, func: impl for<'_q> FnOnce(&'_q Arc<dyn AsRef<[u8]> + Send + Sync>, &'outer_fn mut DependentFontData<'_q>) -> Ret, ) -> Ret

Calls given closure func with an unique reference to dependent.

fn borrow_dependent<'_q>(&'_q self) -> &'_q DependentFontData<'_q>

Borrows dependent.

fn into_owner(self) -> Arc<dyn AsRef<[u8]> + Send + Sync>

Consumes self and returns the owner.

impl FontCell

fn px_scale_factor(&self, scale: f32) -> f32

fn allocate_glyph_uncached( &mut self, atlas: &mut TextureAtlas, metrics: &StyledMetrics, glyph_info: &GlyphInfo, bin: SubpixelBin, location: LocationRef<'_>, ) -> Option<GlyphAllocation>