epaint::text::font

Struct FontFace

pub struct FontFace {
    name: String,
    font: FontCell,
    tweak: FontTweak,
    glyph_info_cache: HashMap<char, GlyphInfo>,
    glyph_alloc_cache: HashMap<GlyphCacheKey, GlyphAllocation>,
}

A specific font face. The interface uses points as the unit for everything.

Fields
name: Stringfont: FontCelltweak: FontTweakglyph_info_cache: HashMap<char, GlyphInfo>glyph_alloc_cache: HashMap<GlyphCacheKey, GlyphAllocation>
Implementations

impl FontFace

pub fn new( options: TextOptions, name: String, font_data: Arc<dyn AsRef<[u8]> + Send + Sync>, index: u32, tweak: FontTweak, ) -> Result<Self, Box<dyn Error>>

fn ignore_character(&self, chr: char) -> bool

Code points that will always be replaced by the replacement character.

See also invisible_char.

fn characters(&self) -> impl Iterator<Item = char> + '_

An un-ordered iterator over all supported characters.

pub(super) fn glyph_info(&mut self, c: char) -> Option<GlyphInfo>

\n will result in None

pub(super) fn pair_kerning_pixels( &self, metrics: &StyledMetrics, last_glyph_id: GlyphId, glyph_id: GlyphId, ) -> f32

pub fn pair_kerning( &self, metrics: &StyledMetrics, last_glyph_id: GlyphId, glyph_id: GlyphId, ) -> f32

pub fn styled_metrics( &self, pixels_per_point: f32, font_size: f32, coords: &VariationCoords, ) -> StyledMetrics

pub fn allocate_glyph( &mut self, atlas: &mut TextureAtlas, metrics: &StyledMetrics, glyph_info: GlyphInfo, chr: char, h_pos: f32, ) -> (GlyphAllocation, i32)