epaint::text::font

Struct FontImpl

pub struct FontImpl {
    name: String,
    ab_glyph_font: FontArc,
    tweak: FontTweak,
    glyph_info_cache: HashMap<char, GlyphInfo>,
    glyph_alloc_cache: HashMap<GlyphCacheKey, GlyphAllocation>,
}

A specific font face. The interface uses points as the unit for everything.

Fields
name: Stringab_glyph_font: FontArctweak: FontTweakglyph_info_cache: HashMap<char, GlyphInfo>glyph_alloc_cache: HashMap<GlyphCacheKey, GlyphAllocation>
Implementations

impl FontImpl

pub fn new(name: String, ab_glyph_font: FontArc, tweak: FontTweak) -> Self

fn ignore_character(&self, chr: char) -> bool

Code points that will always be replaced by the replacement character.

See also invisible_char.

fn characters(&self) -> impl Iterator<Item = char> + '_

An un-ordered iterator over all supported characters.

pub(super) fn glyph_info(&mut self, c: char) -> Option<GlyphInfo>

\n will result in None

pub(super) fn pair_kerning_pixels( &self, metrics: &ScaledMetrics, last_glyph_id: GlyphId, glyph_id: GlyphId, ) -> f32

pub fn pair_kerning( &self, metrics: &ScaledMetrics, last_glyph_id: GlyphId, glyph_id: GlyphId, ) -> f32

pub fn scaled_metrics( &self, pixels_per_point: f32, font_size: f32, ) -> ScaledMetrics

pub fn allocate_glyph( &mut self, atlas: &mut TextureAtlas, metrics: &ScaledMetrics, glyph_info: GlyphInfo, chr: char, h_pos: f32, ) -> (GlyphAllocation, i32)