epaint::text

Module fonts

Structs

CachedFamily 🔒

Cached data for working with a font family (e.g. doing character lookups).

CachedGalley 🔒

FontData

A .ttf or .otf file and a font face index.

FontDefinitions

Describes the font data and the sizes to use.

FontFaceKey 🔒

Unique ID for looking up a single font face/file.

FontId

How to select a sized font.

FontInsert

FontTweak

Extra scale and vertical tweak to apply to all text of a certain font.

Fonts

The collection of fonts used by epaint.

FontsImpl

The collection of fonts used by epaint.

FontsView

The context’s collection of fonts, with this context’s pixels_per_point. This is what you use to do text layout.

GalleyCache 🔒

InsertFontFamily

Enums

FontFamily

Font of unknown size.

FontPriority

Functions

blob_from_font_data 🔒

should_cache_each_paragraph_individually 🔒

If true, lay out and cache each paragraph (sections separated by newlines) individually.

Type Aliases

Blob