Struct epaint::text::FontsImpl

pub struct FontsImpl {
    pixels_per_point: f32,
    max_texture_side: usize,
    definitions: FontDefinitions,
    atlas: Arc<Mutex<TextureAtlas>>,
    font_impl_cache: FontImplCache,
    sized_family: HashMap<(OrderedFloat<f32>, FontFamily), Font>,
}

The collection of fonts used by epaint.

Required in order to paint text.

Fields
pixels_per_point: f32max_texture_side: usizedefinitions: FontDefinitionsatlas: Arc<Mutex<TextureAtlas>>font_impl_cache: FontImplCachesized_family: HashMap<(OrderedFloat<f32>, FontFamily), Font>
Implementations

impl FontsImpl

pub fn new( pixels_per_point: f32, max_texture_side: usize, definitions: FontDefinitions, ) -> Self

Create a new FontsImpl for text layout. This call is expensive, so only create one FontsImpl and then reuse it.

pub fn pixels_per_point(&self) -> f32

pub fn definitions(&self) -> &FontDefinitions

pub fn font(&mut self, font_id: &FontId) -> &mut Font

Get the right font implementation from size and FontFamily.

fn glyph_width(&mut self, font_id: &FontId, c: char) -> f32

Width of this character in points.

pub fn has_glyph(&mut self, font_id: &FontId, c: char) -> bool

Can we display this glyph?

pub fn has_glyphs(&mut self, font_id: &FontId, s: &str) -> bool

Can we display all the glyphs in this text?

fn row_height(&mut self, font_id: &FontId) -> f32

Height of one row of text in points.