epaint::text
Struct Glyph

pub struct Glyph {
    pub chr: char,
    pub pos: Pos2,
    pub advance_width: f32,
    pub line_height: f32,
    pub font_ascent: f32,
    pub font_height: f32,
    pub font_impl_ascent: f32,
    pub font_impl_height: f32,
    pub uv_rect: UvRect,
    pub section_index: u32,
}

Fields
chr: char

The character this glyph represents.

pos: Pos2

Baseline position, relative to the galley. Logical position: pos.y is the same for all chars of the same TextFormat.

advance_width: f32

Logical width of the glyph.

line_height: f32

Height of this row of text.

Usually same as Self::font_height, unless explicitly overridden by TextFormat::line_height.

font_ascent: f32

The ascent of this font.

font_height: f32

The row/line height of this font.

font_impl_ascent: f32

The ascent of the sub-font within the font (FontImpl).

font_impl_height: f32

The row/line height of the sub-font within the font (FontImpl).

uv_rect: UvRect

Position and size of the glyph in the font texture, in texels.

section_index: u32

Index into LayoutJob::sections. Decides color etc.

Implementations

impl Glyph

pub fn size(&self) -> Vec2

pub fn max_x(&self) -> f32

pub fn logical_rect(&self) -> Rect

Same y range for all characters with the same TextFormat.