epaint::text

Struct LayoutJob

pub struct LayoutJob {
    pub text: String,
    pub sections: Vec<LayoutSection>,
    pub wrap: TextWrapping,
    pub first_row_min_height: f32,
    pub break_on_newline: bool,
    pub halign: Align,
    pub justify: bool,
    pub round_output_to_gui: bool,
}

Describes the task of laying out text.

This supports mixing different fonts, color and formats (underline etc).

Pass this to crate::FontsView::layout_job or crate::text::layout.

Example:

use epaint::{Color32, text::{LayoutJob, TextFormat}, FontFamily, FontId};

let mut job = LayoutJob::default();
job.append(
    "Hello ",
    0.0,
    TextFormat {
        font_id: FontId::new(14.0, FontFamily::Proportional),
        color: Color32::WHITE,
        ..Default::default()
    },
);
job.append(
    "World!",
    0.0,
    TextFormat {
        font_id: FontId::new(14.0, FontFamily::Monospace),
        color: Color32::BLACK,
        ..Default::default()
    },
);

As you can see, constructing a LayoutJob is currently a lot of work. It would be nice to have a helper macro for it!

Fields
text: String

The complete text of this job, referenced by LayoutSection.

sections: Vec<LayoutSection>

The different section, which can have different fonts, colors, etc.

wrap: TextWrapping

Controls the text wrapping and elision.

first_row_min_height: f32

The first row must be at least this high. This is in case we lay out text that is the continuation of some earlier text (sharing the same row), in which case this will be the height of the earlier text. In other cases, set this to 0.0.

break_on_newline: bool

If true, all \n characters will result in a new paragraph, starting on a new row.

If false, all \n characters will be ignored and show up as the replacement character.

Default: true.

halign: Align

How to horizontally align the text (Align::LEFT, Align::Center, Align::RIGHT).

justify: bool

Justify text so that word-wrapped rows fill the whole TextWrapping::max_width.

round_output_to_gui: bool

Round output sizes using emath::GuiRounding, to avoid rounding errors in layout code.

Implementations

impl LayoutJob

pub fn simple( text: String, font_id: FontId, color: Color32, wrap_width: f32, ) -> Self

Break on \n and at the given wrap width.

pub fn simple_format(text: String, format: TextFormat) -> Self

Break on \n

pub fn simple_singleline(text: String, font_id: FontId, color: Color32) -> Self

Does not break on \n, but shows the replacement character instead.

pub fn single_section(text: String, format: TextFormat) -> Self

pub fn is_empty(&self) -> bool

pub fn append(&mut self, text: &str, leading_space: f32, format: TextFormat)

Helper for adding a new section when building a LayoutJob.

pub fn font_height(&self, fonts: &mut FontsView<'_>) -> f32

The height of the tallest font used in the job.

Returns a value rounded to emath::GUI_ROUNDING.

pub fn effective_wrap_width(&self) -> f32

The wrap with, with a small margin in some cases.