epaint::text

Module text_layout_types

Structs

Galley

Text that has been laid out, ready for painting.

Glyph

LayoutJob

Describes the task of laying out text.

LayoutSection

PlacedRow

Row

RowVisuals

The tessellated output of a row.

Tag

An OpenType tag.

TextFormat

Formatting option for a section of text.

TextWrapping

Controls the text wrapping and elision of a LayoutJob.

VariationCoords

List of font variation coordinates by axis tag. If more than one coordinate for a given axis is provided, the last one added is used.

Enums

TextWrapMode

How to wrap and elide text.

Traits

IntoTag

Helper trait for all types that can be parsed as a font_types::Tag.