Struct epaint::textures::TextureManager

pub struct TextureManager { /* private fields */ }

Low-level manager for allocating textures.

Communicates with the painting subsystem using Self::take_delta.

Implementations

impl TextureManager

pub fn alloc(
&mut self,
name: String,
image: ImageData,
options: TextureOptions
) -> TextureId

Allocate a new texture.

The given name can be useful for later debugging.

The returned TextureId will be TextureId::Managed, with an index starting from zero and increasing with each call to Self::alloc.

The first texture you allocate will be TextureId::Managed(0) == TexureId::default() and MUST have a white pixel at (0,0) (crate::WHITE_UV).

The texture is given a retain-count of 1, requiring one call to Self::free to free it.

pub fn set(&mut self, id: TextureId, delta: ImageDelta)

Assign a new image to an existing texture, or update a region of it.

pub fn free(&mut self, id: TextureId)

Free an existing texture.

pub fn retain(&mut self, id: TextureId)

Increase the retain-count of the given texture.

For each time you call Self::retain you must call Self::free on additional time.

pub fn take_delta(&mut self) -> TexturesDelta

Take and reset changes since last frame.

These should be applied to the painting subsystem each frame.

pub fn meta(&self, id: TextureId) -> Option<&TextureMeta>

Get meta-data about a specific texture.

pub fn allocated(
&self
) -> impl ExactSizeIterator<Item = (&TextureId, &TextureMeta)>

Get meta-data about all allocated textures in some arbitrary order.

pub fn num_allocated(&self) -> usize

Total number of allocated textures.