Crate epaint

2D graphics/rendering. Fonts, textures, color, geometry, tessellation etc.

Re-exports

pub use color::Color32;

pub use color::Rgba;

pub use stats::PaintStats;

pub use tessellator::TessellationOptions;

pub use tessellator::Tessellator;

pub use ahash;

pub use emath;

Modules

color
Color conversions and types.

mutex
Helper module that wraps some Mutex types with different implementations.

shape_transform

stats
Collect statistics about what is being painted.

tessellator
Converts graphics primitives into textured triangles.

text
Everything related to text, fonts, text layout, cursors etc.

Structs

ClippedMesh
A Mesh within a clip rectangle.

ClippedShape
A Shape within a clip rectangle.

Galley
A collection of text locked into place.

Mesh
Textured triangles in two dimensions.

Mesh16
A version of Mesh that uses 16-bit indices.

Shadow
The color and fuzziness of a fuzzy shape. Can be used for a rectangular shadow with a soft penumbra.

Stroke
Describes the width and color of a line.

Texture
An 8-bit texture containing font data.

TextureAtlas
Contains font data in an atlas, where each character occupied a small rectangle.

Vertex
The 2D vertex type.

Enums

Shape
A paint primitive such as a circle or a piece of text. Coordinates are all screen space points (not physical pixels).

TextStyle
One of a few categories of styles of text, e.g. body, button or heading.

TextureId
What texture to use in a Mesh mesh.

Constants

WHITE_UV
The UV coordinate of a white region of the texture mesh. The default egui texture has the top-left corner pixel fully white. You need need use a clamping texture sampler for this to work (so it doesn't do bilinear blending with bottom right corner).