Crate epaint

A simple 2D graphics library for turning simple 2D shapes and text into textured triangles.

Made for egui.

Create some Shape:s and pass them to tessellate_shapes to generate Mesh:es that you can then paint using some graphics API of your choice (e.g. OpenGL).

Feature flags

Re-exports

pub use image::ColorImage;

pub use image::FontImage;

pub use image::ImageData;

pub use image::ImageDelta;

pub use stats::PaintStats;

pub use tessellator::tessellate_shapes;

pub use tessellator::TessellationOptions;

pub use tessellator::Tessellator;

pub use text::FontFamily;

pub use text::FontId;

pub use text::Fonts;

pub use text::Galley;

pub use textures::TextureManager;

pub use ahash;

pub use ecolor;

pub use emath;

Modules

bezier 🔒

image

mesh 🔒

mutex

Helper module that wraps some Mutex types with different implementations.

shadow 🔒

shape 🔒

The different shapes that can be painted.

shape_transform

stats

Collect statistics about what is being painted.

stroke 🔒

tessellator

Converts graphics primitives into textured triangles.

text

Everything related to text, fonts, text layout, cursors etc.

texture_atlas 🔒

texture_handle 🔒

textures

util

Macros

epaint_assert

An assert that is only active when epaint is compiled with the extra_asserts feature or with the extra_debug_asserts feature in debug builds.

Structs

CircleShape

How to paint a circle.

ClippedPrimitive

A Mesh or PaintCallback within a clip rectangle.

ClippedShape

A Shape within a clip rectangle.

Color32

This format is used for space-efficient color representation (32 bits).

CubicBezierShape

A cubic Bézier Curve.

Hsva

Hue, saturation, value, alpha. All in the range [0, 1]. No premultiplied alpha.

HsvaGamma

Like Hsva but with the v value (brightness) being gamma corrected so that it is somewhat perceptually even.

Mesh

Textured triangles in two dimensions.

Mesh16

A version of Mesh that uses 16-bit indices.

PaintCallback

If you want to paint some 3D shapes inside an egui region, you can use this.

PaintCallbackInfo

Information passed along with PaintCallback (Shape::Callback).

PathShape

A path which can be stroked and/or filled (if closed).

Pos2

A position on screen.

QuadraticBezierShape

A quadratic Bézier Curve.

Rect

A rectangular region of space.

RectShape

How to paint a rectangle.

Rgba

0-1 linear space RGBA color with premultiplied alpha.

Rounding

How rounded the corners of things should be

Shadow

The color and fuzziness of a fuzzy shape. Can be used for a rectangular shadow with a soft penumbra.

Stroke

Describes the width and color of a line.

TextShape

How to paint some text on screen.

TextureAtlas

Contains font data in an atlas, where each character occupied a small rectangle.

TextureHandle

Used to paint images.

Vec2

A vector has a direction and length. A Vec2 is often used to represent a size.

Vertex

The 2D vertex type.

Enums

Primitive

A rendering primitive - either a Mesh or a PaintCallback.

Shape

A paint primitive such as a circle or a piece of text. Coordinates are all screen space points (not physical pixels).

TextureId

What texture to use in a Mesh mesh.

Constants

WHITE_UV

The UV coordinate of a white region of the texture mesh. The default egui texture has the top-left corner pixel fully white. You need need use a clamping texture sampler for this to work (so it doesn’t do bilinear blending with bottom right corner).

Functions

f32_hash 🔒

f64_hash 🔒

pos2

pos2(x, y) == Pos2::new(x, y)

vec2

vec2(x, y) == Vec2::new(x, y)