Struct epaint::color::Color32

pub struct Color32(_);

This format is used for space-efficient color representation (32 bits).

Instead of manipulating this directly it is often better to first convert it to either Rgba or Hsva.

Internally this uses 0-255 gamma space sRGBA color with premultiplied alpha. Alpha channel is in linear space.

Implementations

impl Color32

pub const TRANSPARENT: Color32

pub const BLACK: Color32

pub const LIGHT_GRAY: Color32

pub const GRAY: Color32

pub const WHITE: Color32

pub const RED: Color32

pub const YELLOW: Color32

pub const GREEN: Color32

pub const BLUE: Color32

pub const LIGHT_BLUE: Color32

pub const GOLD: Color32

pub const fn from_rgb(r: u8, g: u8, b: u8) -> Self

pub const fn from_rgb_additive(r: u8, g: u8, b: u8) -> Self

pub const fn from_rgba_premultiplied(r: u8, g: u8, b: u8, a: u8) -> Self

From sRGBA with premultiplied alpha.

pub fn from_rgba_unmultiplied(r: u8, g: u8, b: u8, a: u8) -> Self

From sRGBA WITHOUT premultiplied alpha.

pub const fn new(r: u8, g: u8, b: u8, a: u8) -> Self

👎 Deprecated:
Use from_rgb(..), from_rgba_premultiplied(..) or from_srgba_unmultiplied(..)

pub const fn from_gray(l: u8) -> Self

pub const fn from_black_alpha(a: u8) -> Self

pub fn from_white_alpha(a: u8) -> Self

pub const fn from_additive_luminance(l: u8) -> Self

pub fn is_opaque(&self) -> bool

pub fn r(&self) -> u8

pub fn g(&self) -> u8

pub fn b(&self) -> u8

pub fn a(&self) -> u8

pub fn to_opaque(self) -> Self

Returns an opaque version of self

pub fn additive(self) -> Self

Returns an additive version of self

pub fn to_array(&self) -> [u8; 4]

Premultiplied RGBA

pub fn to_tuple(&self) -> (u8, u8, u8, u8)

Premultiplied RGBA

pub fn linear_multiply(self, factor: f32) -> Color32

Multiply with 0.5 to make color half as opaque.