Enum epaint::Shape

#[must_use = "Add a Shape to a Painter"]pub enum Shape {
    Noop,
    Vec(Vec<Shape>),
    Circle {
        center: Pos2,
        radius: f32,
        fill: Color32,
        stroke: Stroke,
    },
    LineSegment {
        points: [Pos2; 2],
        stroke: Stroke,
    },
    Path {
        points: Vec<Pos2>,
        closed: bool,
        fill: Color32,
        stroke: Stroke,
    },
    Rect {
        rect: Rect,
        corner_radius: f32,
        fill: Color32,
        stroke: Stroke,
    },
    Text {
        pos: Pos2,
        galley: Arc<Galley>,
        color: Color32,
        fake_italics: bool,
    },
    Mesh(Mesh),
}

A paint primitive such as a circle or a piece of text. Coordinates are all screen space points (not physical pixels).

Variants

Noop

Paint nothing. This can be useful as a placeholder.

Vec(Vec<Shape>)

Recursively nest more shapes - sometimes a convenience to be able to do. For performance reasons it is better to avoid it.

Circle

Fields of Circle

center: Pos2radius: f32fill: Color32stroke: Stroke

LineSegment

Fields of LineSegment

points: [Pos2; 2]stroke: Stroke

Path

Fields of Path

points: Vec<Pos2>closed: bool

If true, connect the first and last of the points together. This is required if fill != TRANSPARENT.

fill: Color32stroke: Stroke

Rect

Fields of Rect

rect: Rectcorner_radius: f32

How rounded the corners are. Use 0.0 for no rounding.

fill: Color32stroke: Stroke

Text

Fields of Text

pos: Pos2

Top left corner of the first character..

galley: Arc<Galley>

The layed out text.

color: Color32

Text color (foreground).

fake_italics: bool

If true, tilt the letters for a hacky italics effect.

Mesh(Mesh)

Implementations

impl Shape

Constructors

pub fn line_segment(points: [Pos2; 2], stroke: impl Into<Stroke>) -> Self

pub fn line(points: Vec<Pos2>, stroke: impl Into<Stroke>) -> Self

pub fn closed_line(points: Vec<Pos2>, stroke: impl Into<Stroke>) -> Self

pub fn polygon(
points: Vec<Pos2>,
fill: impl Into<Color32>,
stroke: impl Into<Stroke>
) -> Self

pub fn circle_filled(
center: Pos2,
radius: f32,
fill_color: impl Into<Color32>
) -> Self

pub fn circle_stroke(
center: Pos2,
radius: f32,
stroke: impl Into<Stroke>
) -> Self

pub fn rect_filled(
rect: Rect,
corner_radius: f32,
fill_color: impl Into<Color32>
) -> Self

pub fn rect_stroke(
rect: Rect,
corner_radius: f32,
stroke: impl Into<Stroke>
) -> Self

pub fn text(
fonts: &Fonts,
pos: Pos2,
anchor: Align2,
text: impl Into<String>,
text_style: TextStyle,
color: Color32
) -> Self

impl Shape

Operations

pub fn mesh(mesh: Mesh) -> Self

pub fn triangles(mesh: Mesh) -> Self

👎 Deprecated:
Renamed mesh

pub fn texture_id(&self) -> TextureId

pub fn translate(&mut self, delta: Vec2)

Translate location by this much, in-place