Enum epaint::Shape

pub enum Shape {
    Noop,
    Vec(Vec<Shape>),
    Circle {
        center: Pos2,
        radius: f32,
        fill: Color32,
        stroke: Stroke,
    },
    LineSegment {
        points: [Pos2; 2],
        stroke: Stroke,
    },
    Path {
        points: Vec<Pos2>,
        closed: bool,
        fill: Color32,
        stroke: Stroke,
    },
    Rect {
        rect: Rect,
        corner_radius: f32,
        fill: Color32,
        stroke: Stroke,
    },
    Text {
        pos: Pos2,
        galley: Arc<Galley>,
        color: Color32,
        fake_italics: bool,
    },
    Mesh(Mesh),
}

A paint primitive such as a circle or a piece of text. Coordinates are all screen space points (not physical pixels).

Variants

Noop

Paint nothing. This can be useful as a placeholder.

Vec(Vec<Shape>)

Tuple Fields

0: Vec<Shape>

Recursively nest more shapes - sometimes a convenience to be able to do. For performance reasons it is better to avoid it.

Circle

Fields

center: Pos2

radius: f32

fill: Color32

stroke: Stroke

LineSegment

Fields

points: [Pos2; 2]

stroke: Stroke

Path

Fields

points: Vec<Pos2>

closed: bool

If true, connect the first and last of the points together. This is required if fill != TRANSPARENT.

fill: Color32

Fill is only supported for convex polygons.

stroke: Stroke

Rect

Fields

rect: Rect

corner_radius: f32

How rounded the corners are. Use 0.0 for no rounding.

fill: Color32

stroke: Stroke

Text

Fields

pos: Pos2

Top left corner of the first character..

galley: Arc<Galley>

The layed out text.

color: Color32

Text color (foreground).

fake_italics: bool

If true, tilt the letters for a hacky italics effect.

Mesh(Mesh)

Tuple Fields

0: Mesh

Implementations

impl Shape

Constructors

pub fn line_segment(points: [Pos2; 2], stroke: impl Into<Stroke>) -> Self

A line between two points. More efficient than calling Self::line.

pub fn line(points: Vec<Pos2>, stroke: impl Into<Stroke>) -> Self

A line through many points.

Use Self::line_segment instead if your line only connect two points.

pub fn closed_line(points: Vec<Pos2>, stroke: impl Into<Stroke>) -> Self

A line that closes back to the start point again.

pub fn convex_polygon(
points: Vec<Pos2>,
fill: impl Into<Color32>,
stroke: impl Into<Stroke>
) -> Self

A convex polygon with a fill and optional stroke.

pub fn polygon(
points: Vec<Pos2>,
fill: impl Into<Color32>,
stroke: impl Into<Stroke>
) -> Self

👎 Deprecated:
Renamed convex_polygon

pub fn circle_filled(
center: Pos2,
radius: f32,
fill_color: impl Into<Color32>
) -> Self

pub fn circle_stroke(
center: Pos2,
radius: f32,
stroke: impl Into<Stroke>
) -> Self

pub fn rect_filled(
rect: Rect,
corner_radius: f32,
fill_color: impl Into<Color32>
) -> Self

pub fn rect_stroke(
rect: Rect,
corner_radius: f32,
stroke: impl Into<Stroke>
) -> Self

pub fn text(
fonts: &Fonts,
pos: Pos2,
anchor: Align2,
text: impl ToString,
text_style: TextStyle,
color: Color32
) -> Self

impl Shape

Operations

pub fn mesh(mesh: Mesh) -> Self

pub fn triangles(mesh: Mesh) -> Self

👎 Deprecated:
Renamed mesh

pub fn texture_id(&self) -> TextureId

pub fn translate(&mut self, delta: Vec2)

Translate location by this much, in-place