epaint
Struct QuadraticBezierShape

pub struct QuadraticBezierShape {
    pub points: [Pos2; 3],
    pub closed: bool,
    pub fill: Color32,
    pub stroke: PathStroke,
}

A quadratic Bézier Curve.

See also CubicBezierShape.

Fields
points: [Pos2; 3]

The first point is the starting point and the last one is the ending point of the curve. The middle point is the control points.

closed: boolfill: Color32stroke: PathStroke
Implementations

impl QuadraticBezierShape

pub fn from_points_stroke( points: [Pos2; 3], closed: bool, fill: Color32, stroke: impl Into<PathStroke>, ) -> Self

Create a new quadratic Bézier shape based on the 3 points and stroke.

The first point is the starting point and the last one is the ending point of the curve. The middle point is the control points. The points should be in the order [start, control, end]

pub fn transform(&self, transform: &RectTransform) -> Self

Transform the curve with the given transform.

pub fn to_path_shape(&self, tolerance: Option<f32>) -> PathShape

Convert the quadratic Bézier curve to one PathShape. The tolerance will be used to control the max distance between the curve and the base line.

pub fn visual_bounding_rect(&self) -> Rect

The visual bounding rectangle (includes stroke width)

pub fn logical_bounding_rect(&self) -> Rect

Logical bounding rectangle (ignoring stroke width)

pub fn sample(&self, t: f32) -> Pos2

Calculate the point (x,y) at t based on the quadratic Bézier curve equation. t is in [0.0,1.0] Bézier Curve

pub fn flatten(&self, tolerance: Option<f32>) -> Vec<Pos2>

find a set of points that approximate the quadratic Bézier curve. the number of points is determined by the tolerance. the points may not be evenly distributed in the range [0.0,1.0] (t value)

pub fn for_each_flattened_with_t<F>(&self, tolerance: f32, callback: &mut F)
where F: FnMut(Pos2, f32),

Compute a flattened approximation of the curve, invoking a callback at each step.

The callback takes the point and corresponding curve parameter at each step.

This implements the algorithm described by Raph Levien at https://raphlinus.github.io/graphics/curves/2019/12/23/flatten-quadbez.html