epaint

Module tessellator

Converts graphics primitives into textured triangles.

This module converts lines, circles, text and more represented by Shape into textured triangles represented by Mesh.

Modules

path

Helpers for constructing paths

precomputed_vertices 🔒

Structs

Path

A connected line (without thickness or gaps) which can be tessellated to either to a stroke (with thickness) or a filled convex area. Used as a scratch-pad during tessellation.

PathPoint 🔒

TessellationOptions

Tessellation quality options

Tessellator

Converts Shapes into triangles (Mesh).

Enums

PathType

Functions

cw_signed_area 🔒

fill_closed_path 🔒

Tessellate the given convex area into a polygon.

fill_closed_path_with_uv 🔒

Like fill_closed_path but with texturing.

mul_color 🔒

stroke_and_fill_path 🔒

Tessellate the given path as a stroke with thickness, with optional fill color.

stroke_path 🔒

Tessellate the given path as a stroke with thickness.