Function epaint::tessellator::tessellate_shapes

pub fn tessellate_shapes(
    pixels_per_point: f32,
    options: TessellationOptions,
    font_tex_size: [usize; 2],
    prepared_discs: Vec<PreparedDisc>,
    shapes: Vec<ClippedShape>
) -> Vec<ClippedPrimitive>

Turns Shape:s into sets of triangles.

The given shapes will tessellated in the same order as they are given. They will be batched together by clip rectangle.

pixels_per_point: number of physical pixels to each logical point

options: tessellation quality

shapes: what to tessellate

font_tex_size: size of the font texture. Required to normalize glyph uv rectangles when tessellating text.

prepared_discs: What TextureAtlas::prepared_discs returns. Can safely be set to an empty vec.

The implementation uses a Tessellator.

Returns

A list of clip rectangles with matching Mesh.