Module epaint::text

Everything related to text, fonts, text layout, cursors etc.

Modules

cursor

Different types of text cursors, i.e. ways to point into a super::Galley.

Structs

FontData

A .ttf or .otf file and a font face index.

FontDefinitions

Describes the font data and the sizes to use.

FontId

How to select a sized font.

FontTweak

Extra scale and vertical tweak to apply to all text of a certain font.

Fonts

The collection of fonts used by epaint.

FontsImpl

The collection of fonts used by epaint.

Galley

Text that has been laid out, ready for painting.

Glyph

LayoutJob

Describes the task of laying out text.

LayoutSection

Row

RowVisuals

The tessellated output of a row.

TextFormat

TextWrapping

Enums

FontFamily

Font of unknown size.

Constants

PASSWORD_REPLACEMENT_CHAR

Suggested character to use to replace those in password text fields.

TAB_SIZE

One \t character is this many spaces wide.

Functions

layout

Layout text into a Galley.