Module epaint::text

Everything related to text, fonts, text layout, cursors etc.

Modules

cursor

Different types of text cursors, i.e. ways to point into a super::Galley.

Structs

FontDefinitions

Describes the font data and the sizes to use.

Fonts

The collection of fonts used by epaint.

Galley

A collection of text locked into place.

Row

A typeset piece of text on a single row.

Enums

FontFamily

Which style of font: Monospace or Proportional.

TextStyle

One of a few categories of styles of text, e.g. body, button or heading.

Constants

PASSWORD_REPLACEMENT_CHAR

Suggested character to use to replace those in password text fields.

TAB_SIZE

One \t character is this many spaces wide.