Module epaint::text::cursor

Different types of text cursors, i.e. ways to point into a super::Galley.

Structs

CCursor

Character cursor

Cursor

All different types of cursors together. They all point to the same place, but in their own different ways. pcursor/rcursor can also point to after the end of the paragraph/row. Does not implement PartialEq because you must think which cursor should be equivalent.

PCursor

Paragraph Cursor

RCursor

Row Cursor