epaint::text::font

Struct Font

pub struct Font<'a> {
    pub(super) fonts_by_id: &'a mut IntMap<FontFaceKey, FontImpl>,
    pub(super) cached_family: &'a mut CachedFamily,
    pub(super) atlas: &'a mut TextureAtlas,
}

Wrapper over multiple FontImpl (e.g. a primary + fallbacks for emojis)

Fields
fonts_by_id: &'a mut IntMap<FontFaceKey, FontImpl>cached_family: &'a mut CachedFamilyatlas: &'a mut TextureAtlas
Implementations

impl Font<'_>

pub fn preload_characters(&mut self, s: &str)

pub fn characters(&mut self) -> &BTreeMap<char, Vec<String>>

All supported characters, and in which font they are available in.

pub fn scaled_metrics( &self, pixels_per_point: f32, font_size: f32, ) -> ScaledMetrics

pub fn glyph_width(&mut self, c: char, font_size: f32) -> f32

Width of this character in points.

pub fn has_glyph(&mut self, c: char) -> bool

Can we display this glyph?

pub fn has_glyphs(&mut self, s: &str) -> bool

Can we display all the glyphs in this text?

pub(crate) fn glyph_info(&mut self, c: char) -> (FontFaceKey, GlyphInfo)

\n will (intentionally) show up as the replacement character.