Struct epaint::text::font::Font

pub struct Font {
    fonts: Vec<Arc<FontImpl>>,
    characters: Option<BTreeMap<char, Vec<String>>>,
    replacement_glyph: (usize, GlyphInfo),
    pixels_per_point: f32,
    row_height: f32,
    glyph_info_cache: HashMap<char, (usize, GlyphInfo)>,
}

Wrapper over multiple FontImpl (e.g. a primary + fallbacks for emojis)

Fields
fonts: Vec<Arc<FontImpl>>characters: Option<BTreeMap<char, Vec<String>>>

Lazily calculated.

replacement_glyph: (usize, GlyphInfo)pixels_per_point: f32row_height: f32glyph_info_cache: HashMap<char, (usize, GlyphInfo)>
Implementations

impl Font

pub fn new(fonts: Vec<Arc<FontImpl>>) -> Self

pub fn preload_characters(&mut self, s: &str)

pub fn preload_common_characters(&mut self)

pub fn characters(&mut self) -> &BTreeMap<char, Vec<String>>

All supported characters, and in which font they are available in.

pub fn round_to_pixel(&self, point: f32) -> f32

pub fn row_height(&self) -> f32

Height of one row of text. In points

pub fn uv_rect(&self, c: char) -> UvRect

pub fn glyph_width(&mut self, c: char) -> f32

Width of this character in points.

pub fn has_glyph(&mut self, c: char) -> bool

Can we display this glyph?

pub fn has_glyphs(&mut self, s: &str) -> bool

Can we display all the glyphs in this text?

fn glyph_info(&mut self, c: char) -> (usize, GlyphInfo)

\n will (intentionally) show up as the replacement character.

pub(crate) fn font_impl_and_glyph_info( &mut self, c: char, ) -> (Option<&FontImpl>, GlyphInfo)

pub(crate) fn ascent(&self) -> f32

fn glyph_info_no_cache_or_fallback( &mut self, c: char, ) -> Option<(usize, GlyphInfo)>