epaint::text

Module fonts

Structs

CachedGalley 🔒

FontData

A .ttf or .otf file and a font face index.

FontDefinitions

Describes the font data and the sizes to use.

FontId

How to select a sized font.

FontImplCache 🔒

FontTweak

Extra scale and vertical tweak to apply to all text of a certain font.

Fonts

The collection of fonts used by epaint.

FontsAndCache

FontsImpl

The collection of fonts used by epaint.

GalleyCache 🔒

HashableF32 🔒

Enums

FontFamily

Font of unknown size.

Functions

ab_glyph_font_from_font_data 🔒