Struct epaint::text::Fonts

pub struct Fonts { /* fields omitted */ }

The collection of fonts used by epaint.

Implementations

impl Fonts

pub fn from_definitions(
pixels_per_point: f32,
definitions: FontDefinitions
) -> Self

pub fn pixels_per_point(&self) -> f32

pub fn definitions(&self) -> &FontDefinitions

pub fn round_to_pixel(&self, point: f32) -> f32

pub fn texture(&self) -> Arc<Texture>

Call each frame to get the latest available font texture data.

pub fn glyph_width(&self, text_style: TextStyle, c: char) -> f32

Width of this character in points.

pub fn row_height(&self, text_style: TextStyle) -> f32

Height of one row of text. In points

pub fn layout_no_wrap(&self, text_style: TextStyle, text: String) -> Arc<Galley>

Always returns at least one row. Will line break at \n.

pub fn layout_single_line(
&self,
text_style: TextStyle,
text: String
) -> Arc<Galley>

Typeset the given text onto one row. Any \n will show up as the replacement character. Always returns exactly one Row in the Galley.

Most often you probably want \n to produce a new row, and so Self::layout_no_wrap may be a better choice.

pub fn layout_multiline(
&self,
text_style: TextStyle,
text: String,
max_width_in_points: f32
) -> Arc<Galley>

Always returns at least one row. Will wrap text at the given width.

pub fn layout_multiline_with_indentation_and_max_width(
&self,
text_style: TextStyle,
text: String,
first_row_indentation: f32,
max_width_in_points: f32
) -> Arc<Galley>

first_row_indentation: extra space before the very first character (in points).

max_width_in_points: wrapping width. Always returns at least one row.

pub fn num_galleys_in_cache(&self) -> usize

pub fn end_frame(&self)

Must be called once per frame to clear the Galley cache.