epaint::text
Struct FontsImpl

pub struct FontsImpl { /* private fields */ }

The collection of fonts used by epaint.

Required in order to paint text.

Implementations

impl FontsImpl

pub fn new( pixels_per_point: f32, max_texture_side: usize, definitions: FontDefinitions, ) -> Self

Create a new FontsImpl for text layout. This call is expensive, so only create one FontsImpl and then reuse it.

pub fn pixels_per_point(&self) -> f32

pub fn definitions(&self) -> &FontDefinitions

pub fn font(&mut self, font_id: &FontId) -> &mut Font

Get the right font implementation from size and FontFamily.

pub fn has_glyph(&mut self, font_id: &FontId, c: char) -> bool

Can we display this glyph?

pub fn has_glyphs(&mut self, font_id: &FontId, s: &str) -> bool

Can we display all the glyphs in this text?