Module epaint::text::text_layout

Structs

FormatSummary 🔒

Paragraph 🔒

Temporary storage before line-wrapping.

PointScale 🔒

Represents GUI scale and convenience methods for rounding to pixels.

RowBreakCandidates 🔒

Keeps track of good places to break a long row of text. Will focus primarily on spaces, secondarily on things like -

Functions

add_hline 🔒

add_row_backgrounds 🔒

Create background for glyphs that have them. Creates as few rectangular regions as possible.

add_row_hline 🔒

Add a horizontal line over a row of glyphs with a stroke and y decided by a callback.

format_summary 🔒

galley_from_rows 🔒

Calculate the Y positions and tessellate the text.

halign_and_justify_row 🔒

is_cjk 🔒

is_cjk_break_allowed 🔒

is_cjk_ideograph 🔒

is_kana 🔒

layout

Layout text into a Galley.

layout_section 🔒

line_break 🔒

rect_from_x_range 🔒

We ignore y at this stage

replace_last_glyph_with_overflow_character 🔒

rows_from_paragraphs 🔒

tessellate_glyphs 🔒

tessellate_row 🔒