epaint::text

Module text_layout_types

Structs

Galley

Text that has been laid out, ready for painting.

Glyph

LayoutJob

Describes the task of laying out text.

LayoutSection

PlacedRow

Row

RowVisuals

The tessellated output of a row.

TextFormat

Formatting option for a section of text.

TextWrapping

Controls the text wrapping and elision of a LayoutJob.

Enums

TextWrapMode

How to wrap and elide text.